+++ src/com/ya/grupp5b/skirace/tools/Input.java
@@ -1,7 +1,11 @@
 package com.ya.grupp5b.skirace.tools;
 
 import java.time.LocalTime;
+import java.util.Collections;
 import java.util.Scanner;
+
+import com.ya.grupp5b.skirace.race.Race;
+import com.ya.grupp5b.skirace.skier.CompareTempTime;
 
 public class Input {
 
@@ -41,7 +45,6 @@ public class Input {
 
 	public static String readString() {
 		while (!scan.hasNextLine()) { // Försök fixa så man inte kan skriva in siffror
-//			scan.nextLine();
 			System.out.println("Felaktig inmatning \nProva igen: ");
 		}
 		return scan.nextLine();
@@ -49,7 +52,6 @@ public class Input {
 	}
 
 	// Skapa en metod som kan läsa klockslag
-
 	public static int readInt() {
 		while (!scan.hasNextInt()) {
 			scan.nextLine();
@@ -58,6 +60,47 @@ public class Input {
 		int chosenInt = scan.nextInt();
 		scan.nextLine();
 		return chosenInt;
+	}
+	
+	// METOD: val 3 i menyn - dvs. checka specific tävlandes löpta tid & placering
+	public static void checkTempTime(Race race) {
+		CompareTempTime tempCheck = new CompareTempTime(); // Jämför mellantid
+		Collections.sort(race.getSkierList(), tempCheck);
+
+		System.out.println("\nVilken tävlande vill du kolla nuvarande mellantid & placering för?");
+		int chosenStartNum = Input.readInt();
+		int currentIndex = 0;
+
+		// Hitta <index> för den önskade tävlande, via det valda start nr
+		for (var skier : race.getSkierList()) {
+			if (skier.getSkierNumber() == chosenStartNum) {
+				currentIndex = race.getSkierList().indexOf(skier);
+			}
+		}
+
+		// Kom åt starttid för objektet på det funna indexet
+		LocalTime currentTempTime = Timing.calcDuration(race.getSkierList().get(currentIndex).getIndivStartTime());
+		Output.printChosenTempTime(currentIndex, currentTempTime, race);
+	}
+	
+	// METOD: val 2 i menyn - dvs. registrera måltid för vald åkare
+	public static void regGoalTime(Race race) {
+
+		System.out.println("\nVilken tävlande vill du registrera måltid för? ");
+		int chosenStartNum = Input.readInt();
+		int currentIndex = 0;
+
+		// Hitta <index> för den önskade tävlande, via det valda start nr
+		for (var skier : race.getSkierList()) {
+			if (skier.getSkierNumber() == chosenStartNum) {
+				currentIndex = race.getSkierList().indexOf(skier);
+			}
+		}
+
+		race.getSkierList().get(currentIndex)
+				.setGoalTime(Timing.calcDuration(race.getSkierList().get(currentIndex).getIndivStartTime()));
+
+		Output.printChosenGoalTime(currentIndex, race);
 	}
 
 }
